// src/types.ts
import type { Context } from "hono";
import { z } from "zod";

export type AppContext = Context<{ Bindings: Env }>;

export const Task = z.object({
  name: z.string().openapi({ example: "lorem" }),
  slug: z.string(),
  description: z.string().optional(),
  completed: z.boolean().default(false),
  due_date: z.iso.date()
});

export type Task = z.infer<typeof Task>;

export const ChatMessage = z.object({
  role: z.enum(["user", "assistant", "system"]),
  content: z.string()
});